fix(apollo/events): ensure movieTitle is sent in allMoviesEvents + add ApolloProvider to layout
